# KTLO: Improve eslint performance