--- a/.eslintignore
+++ b/.eslintignore
@@ -1,0 +1,6 @@
+packages/mds-compliance/tests/la-city-boundary.ts
+packages/mds-policy-author/tests/la-city-boundary.ts
+packages/mds-policy/tests/la-city-boundary.ts
+packages/mds-test-data/district-seven.ts
+packages/mds-test-data/la-city-boundary.ts
+packages/mds-test-data/test-areas/*.ts