--- a/.eslintignore
+++ b/.eslintignore
@@ -1,1 +1,1 @@
-packages/mds-test-data/
+packages/mds-test-data/test-areas/