--- a/.eslintrc.json
+++ b/.eslintrc.json
@@ -4,7 +4,7 @@
     "es6": true,
     "mocha": true
   },
-  "plugins": ["import", "prettier", "@typescript-eslint", "promise"],
+  "plugins": ["import", "prettier", "@typescript-eslint", "promise", "log"],
   "parser": "@typescript-eslint/parser",
   "parserOptions": {
     "ecmaVersion": 2018,