--- a/.eslintrc.json
+++ b/.eslintrc.json
@@ -4,7 +4,14 @@
     "es6": true,
     "mocha": true
   },
-  "plugins": ["import", "prettier", "@typescript-eslint", "promise", "log"],
+  "plugins": [
+    "import",
+    "prettier",
+    "@typescript-eslint",
+    "promise",
+    "log",
+    "progress"
+  ],
   "parser": "@typescript-eslint/parser",
   "parserOptions": {
     "ecmaVersion": 2018,
@@ -49,6 +56,7 @@
     "no-useless-constructor": "off",
     "no-var": "error",
     "prettier/prettier": "warn",
+    "progress/activate": "off",
     "promise/always-return": "error",
     "promise/avoid-new": "off",
     "promise/catch-or-return": "error",