--- a/.huskyrc.json
+++ b/.huskyrc.json
@@ -1,5 +1,0 @@
-{
-  "hooks": {
-    "pre-commit": "pre-commit"
-  }
-}