--- a/.lintstagedrc.json
+++ b/.lintstagedrc.json
@@ -1,0 +1,4 @@
+{
+  "*.ts": "pnpm lint:eslint",
+  "package.json": "pnpm lint:prettier"
+}