--- a/@types/la-city-boundary.d.ts
+++ b/@types/la-city-boundary.d.ts
@@ -1,1 +1,0 @@
-declare module 'la_city_boundary'