--- a/Jenkinsfile
+++ b/Jenkinsfile
@@ -53,7 +53,8 @@ pipeline {
             trap cleanup EXIT
 
             pnpm clean
-            PG_NAME=postgres PG_HOST=localhost PG_USER=postgres REDIS_HOST=localhost pnpm lint && pnpm test
+            pnpm lint
+            PG_NAME=postgres PG_HOST=localhost PG_USER=postgres REDIS_HOST=localhost pnpm test
           '''
         }
       }