--- a/container-images/mds-agency/tsconfig.eslint.json
+++ b/container-images/mds-agency/tsconfig.eslint.json
@@ -1,1 +1,0 @@
-{ "extends": "../../tsconfig.eslint.json" }