--- a/container-images/mds-compliance/tsconfig.eslint.json
+++ b/container-images/mds-compliance/tsconfig.eslint.json
@@ -1,1 +1,0 @@
-{ "extends": "../../tsconfig.eslint.json" }