--- a/container-images/mds-daily/tsconfig.eslint.json
+++ b/container-images/mds-daily/tsconfig.eslint.json
@@ -1,1 +1,0 @@
-{ "extends": "../../tsconfig.eslint.json" }