--- a/container-images/mds-geography-service/package.json
+++ b/container-images/mds-geography-service/package.json
@@ -18,9 +18,7 @@
     "bundle:production": "webpack --mode=production --env npm_package_name=${npm_package_name} --env npm_package_version=${npm_package_version}",
     "image": "pnpm image:production",
     "image:development": "pnpm bundle:development && ../../bin/build-helper buildImage mds-geography-service:${npm_package_version}",
-    "image:production": "pnpm bundle:production && ../../bin/build-helper buildImage mds-geography-service:${npm_package_version}",
-    "test": "pnpm test:eslint",
-    "test:eslint": "eslint --ignore-path ../../.gitignore '**/*.ts'"
+    "image:production": "pnpm bundle:production && ../../bin/build-helper buildImage mds-geography-service:${npm_package_version}"
   },
   "dependencies": {
     "@mds-core/mds-geography-service": "0.1.0",