--- a/container-images/mds-geography/tsconfig.eslint.json
+++ b/container-images/mds-geography/tsconfig.eslint.json
@@ -1,1 +1,0 @@
-{ "extends": "../../tsconfig.eslint.json" }