--- a/container-images/mds-ingest-service/tsconfig.eslint.json
+++ b/container-images/mds-ingest-service/tsconfig.eslint.json
@@ -1,1 +1,0 @@
-{ "extends": "../../tsconfig.eslint.json" }