--- a/container-images/mds-jurisdiction/tsconfig.eslint.json
+++ b/container-images/mds-jurisdiction/tsconfig.eslint.json
@@ -1,1 +1,0 @@
-{ "extends": "../../tsconfig.eslint.json" }