--- a/container-images/mds-metrics-sheet/tsconfig.eslint.json
+++ b/container-images/mds-metrics-sheet/tsconfig.eslint.json
@@ -1,1 +1,0 @@
-{ "extends": "../../tsconfig.eslint.json" }