--- a/container-images/mds-policy-author/tsconfig.eslint.json
+++ b/container-images/mds-policy-author/tsconfig.eslint.json
@@ -1,1 +1,0 @@
-{ "extends": "../../tsconfig.eslint.json" }