--- a/container-images/mds-policy/tsconfig.eslint.json
+++ b/container-images/mds-policy/tsconfig.eslint.json
@@ -1,1 +1,0 @@
-{ "extends": "../../tsconfig.eslint.json" }