--- a/container-images/mds-transactions-api/tsconfig.eslint.json
+++ b/container-images/mds-transactions-api/tsconfig.eslint.json
@@ -1,1 +1,0 @@
-{ "extends": "../../tsconfig.eslint.json" }