--- a/container-images/mds-transactions-service/package.json
+++ b/container-images/mds-transactions-service/package.json
@@ -14,9 +14,7 @@
   ],
   "scripts": {
     "bundle": "webpack --mode=production --env npm_package_name=${npm_package_name} --env npm_package_version=${npm_package_version}",
-    "image": "pnpm bundle && ../../bin/build-helper buildImage mds-transactions-service:${npm_package_version}",
-    "test": "pnpm test:eslint",
-    "test:eslint": "eslint --ignore-path ../../.gitignore '**/*.ts'"
+    "image": "pnpm bundle && ../../bin/build-helper buildImage mds-transactions-service:${npm_package_version}"
   },
   "dependencies": {
     "@mds-core/mds-transactions-service": "0.1.0",