--- a/package.json
+++ b/package.json
@@ -20,7 +20,7 @@
     "push": "bin/build-helper pushImages",
     "run:concurrent": "pnpm recursive run --filter=!root",
     "run:sequential": "pnpm recursive run --workspace-concurrency=1 --filter=!root",
-    "pretest": "pnpm-ci audit -i 1556,1561 --strict && pnpm lint",
+    "pretest": "pnpm-ci audit -i 1556,1561 --strict",
     "test": "pnpm run:sequential test",
     "values": "bin/build-helper templateValues"
   },