--- a/package.json
+++ b/package.json
@@ -42,6 +42,7 @@
     "eslint-plugin-import": "2.22.1",
     "eslint-plugin-log": "1.2.7",
     "eslint-plugin-prettier": "3.3.1",
+    "eslint-plugin-progress": "0.0.1",
     "eslint-plugin-promise": "4.2.1",
     "git-revision-webpack-plugin": "3.0.6",
     "husky": "5.0.9",