--- a/packages/mds-agency/tsconfig.eslint.json
+++ b/packages/mds-agency/tsconfig.eslint.json
@@ -1,1 +1,0 @@
-{ "extends": "../../tsconfig.eslint.json" }