--- a/packages/mds-api-authorizer/package.json
+++ b/packages/mds-api-authorizer/package.json
@@ -14,9 +14,7 @@
   ],
   "scripts": {
     "build": "tsc --build tsconfig.build.json",
-    "test": "pnpm test:eslint && pnpm test:unit",
-    "test:eslint": "eslint --ignore-path ../../.gitignore '**/*.ts'",
-    "test:unit": "DOTENV_CONFIG_PATH=../../.env nyc --use-spawn-wrap  ts-mocha --project ../../tsconfig.json"
+    "test": "DOTENV_CONFIG_PATH=../../.env nyc --use-spawn-wrap  ts-mocha --project ../../tsconfig.json"
   },
   "dependencies": {
     "@mds-core/mds-test-data": "0.1.26",