--- a/packages/mds-api-authorizer/tsconfig.eslint.json
+++ b/packages/mds-api-authorizer/tsconfig.eslint.json
@@ -1,1 +1,0 @@
-{ "extends": "../../tsconfig.eslint.json" }