--- a/packages/mds-api-helpers/tsconfig.eslint.json
+++ b/packages/mds-api-helpers/tsconfig.eslint.json
@@ -1,1 +1,0 @@
-{ "extends": "../../tsconfig.eslint.json" }