--- a/packages/mds-api-server/tsconfig.eslint.json
+++ b/packages/mds-api-server/tsconfig.eslint.json
@@ -1,1 +1,0 @@
-{ "extends": "../../tsconfig.eslint.json" }