--- a/packages/mds-compliance/tests/engine-test.spec.ts
+++ b/packages/mds-compliance/tests/engine-test.spec.ts
@@ -17,13 +17,11 @@
 import test from 'unit.js'
 import fs from 'fs'
 
-import { makeDevices, makeEventsWithTelemetry } from '@mds-core/mds-test-data'
+import { makeDevices, makeEventsWithTelemetry, LA_CITY_BOUNDARY } from '@mds-core/mds-test-data'
 import { RULE_TYPES, Geography, Policy, Device, VehicleEvent } from '@mds-core/mds-types'
 
-import { FeatureCollection } from 'geojson'
 import { RuntimeError, minutes } from '@mds-core/mds-utils'
 import { ValidationError, validateEvents, validateGeographies, validatePolicies } from '@mds-core/mds-schema-validators'
-import { la_city_boundary } from './la-city-boundary'
 import { processPolicy, getSupersedingPolicies, getRecentEvents } from '../mds-compliance-engine'
 
 let policies: Policy[] = []
@@ -31,9 +29,7 @@ let low_count_policies: Policy[] = []
 
 const CITY_OF_LA = '1f943d59-ccc9-4d91-b6e2-0c5e771cbc49'
 
-const geographies: Geography[] = [
-  { name: 'la', geography_id: CITY_OF_LA, geography_json: la_city_boundary as FeatureCollection }
-]
+const geographies: Geography[] = [{ name: 'la', geography_id: CITY_OF_LA, geography_json: LA_CITY_BOUNDARY }]
 
 process.env.TIMEZONE = 'America/Los_Angeles'