--- a/packages/mds-test-data/test-areas/district-seven.ts
+++ b/packages/mds-test-data/test-areas/district-seven.ts
@@ -1108,3 +1108,4 @@ const DISTRICT_SEVEN: FeatureCollection = {
 }
 
 export { DISTRICT_SEVEN }
+export default DISTRICT_SEVEN