## 📚 Purpose
Improve linting performance by only invoking the linter once and ignoring some very expensive test data files. This also eliminates the need for a separate `tsconfig.eslint.json` file in each package.

***Before***
```
pnpm test:eslint  438.70s user 34.31s system 379% cpu 2:04.64 total
```
***After***
```
pnpm lint  41.95s user 2.40s system 169% cpu 26.133 total
```